refactor(38139): refactor styles
