+++ hivemq-edge-frontend/src/modules/Workspace/components/wizard/WizardSelectionRestrictions.tsx
@@ -8,7 +8,7 @@ import type { ProtocolAdapter } from '@/api/__generated__'
 import { useWizardState } from '@/modules/Workspace/hooks/useWizardStore'
 import { EdgeTypes, IdStubs, NodeTypes } from '@/modules/Workspace/types'
 import type { SelectionConstraints } from './types'
-import { GHOST_COLOR_EDGE, GHOST_EDGE_STYLE } from './utils/ghostNodeFactory'
+import { GHOST_COLOR_EDGE, GHOST_EDGE_STYLE } from './utils/styles'
 import { useProtocolAdaptersContext } from './hooks/useProtocolAdaptersContext'
 import { canNodeBeGrouped } from './utils/groupConstraints'
