+++ hivemq-edge-frontend/src/modules/Workspace/components/wizard/hooks/useCompleteUtilities.ts
@@ -1,10 +1,7 @@
 import { useReactFlow } from '@xyflow/react'
-import {
-  GHOST_SUCCESS_DIMMED_TRANSITION,
-  removeGhostEdges,
-  removeGhostNodes,
-} from '@/modules/Workspace/components/wizard/utils/ghostNodeFactory.ts'
+import { removeGhostEdges, removeGhostNodes } from '@/modules/Workspace/components/wizard/utils/ghostNodeFactory.ts'
 import { useWizardStore } from '@/modules/Workspace/hooks/useWizardStore.ts'
+import { GHOST_SUCCESS_OPACITY_TRANSITION } from '../utils/styles'
 
 export const useCompleteUtilities = () => {
   const { getNodes, setNodes, getEdges, setEdges } = useReactFlow()
@@ -46,7 +43,7 @@ export const useCompleteUtilities = () => {
           style: {
             ...node.style,
             opacity: 0.3,
-            transition: GHOST_SUCCESS_DIMMED_TRANSITION,
+            transition: GHOST_SUCCESS_OPACITY_TRANSITION,
           },
         }
       }
