+++ hivemq-edge-frontend/src/modules/Workspace/components/wizard/utils/ghostNodeFactory.ts
@@ -8,6 +8,7 @@ import { EdgeTypes, IdStubs, NodeTypes } from '@/modules/Workspace/types'
 import { getAutoIncludedNodes } from './groupConstraints'
 
 import i18n from '@/config/i18n.config.ts'
+import { GHOST_COLOR_EDGE, GHOST_EDGE_STYLE, GHOST_STYLE, GHOST_STYLE_ENHANCED, GHOST_STYLE_SELECTABLE } from './styles'
 
 /**
  * Positioning constants (from nodes-utils.ts)
@@ -17,63 +18,12 @@ const MAX_ADAPTERS = 10
 const GLUE_SEPARATOR = 200
 
 /**
- * GHOST-related colors
- * TODO[NVL] These must be integrated in the theme
- */
-export const GHOST_COLOR_BACKGROUND = '#EBF8FF'
-export const GHOST_COLOR_EDGE = '#4299E1'
-export const GHOST_SUCCESS_SHADOW = '0 0 0 4px rgba(72, 187, 120, 0.6), 0 0 20px rgba(72, 187, 120, 0.4)'
-export const GHOST_SUCCESS_TRANSITION = 'box-shadow 0.3s ease-in'
-export const GHOST_SUCCESS_DIMMED_TRANSITION = 'box-shadow 0.5s ease-out'
-
-/**
  * Base ghost node properties
  */
 const GHOST_BASE = {
   draggable: false,
   selectable: false,
   connectable: false,
-}
-
-/**
- * Enhanced ghost node styling with glowing effect
- */
-export const GHOST_STYLE_ENHANCED = {
-  opacity: 0.75,
-  border: `3px dashed ${GHOST_COLOR_EDGE}`,
-  backgroundColor: GHOST_COLOR_BACKGROUND,
-  boxShadow: '0 0 0 4px rgba(66, 153, 225, 0.4), 0 0 20px rgba(66, 153, 225, 0.6)',
-  pointerEvents: 'none' as const,
-  transition: 'all 0.3s ease',
-}
-
-/**
- * Selectable ghost node styling - allows clicking to see edge highlighting
- */
-export const GHOST_STYLE_SELECTABLE = {
-  ...GHOST_STYLE_ENHANCED,
-  pointerEvents: 'all' as const, // Allow interaction
-  cursor: 'pointer' as const,
-}
-
-/**
- * Ghost edge styling
- */
-export const GHOST_EDGE_STYLE = {
-  stroke: GHOST_COLOR_EDGE,
-  strokeWidth: 2,
-  strokeDasharray: '5,5',
-  opacity: 0.6,
-}
-
-/**
- * Legacy ghost node styling (for backward compatibility)
- */
-export const GHOST_STYLE = {
-  opacity: 0.6,
-  border: `2px dashed ${GHOST_COLOR_EDGE}`,
-  backgroundColor: GHOST_COLOR_BACKGROUND,
-  pointerEvents: 'none' as const,
 }
 
 /**
